+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Remove factors from requirements if already implied by other requirement
 list.
+- Added tiling method `is_empty_cell` and `is_monotone_cell`
 ### Fixed
 - Infinite recursion issue in get_genf.
 - Close mongo when finished.